Pattern for nested view models
The crux of it is having @State outside the content view is bad idea.
Instead a ViewModel should contain only reference types, nested views should have their
own ObservableObject ViewModel classes. The nested Views then hold a collection
of @Bindings that can draw themselves given the ViewModel

This leaves the rather large hole of not being able to read in the ViewModal at runtime with Codable
Either I'm doing something wrong or this is a bit of a gap with SwiftUI