Using @State for the viewModel object was wrong
I was mutating the reference. This worked for objects contained in the ViewModel but primitive types
were not getting updated. Not sure on why that is, SwiftUI must bind value types to the view and references
for reference types? Anyway

Correct thing is to use a StateObject but that raises a big problem: I cannot assign to it.
This means I can't read a Codable object from disk when the view appear, Codable
only returns new instances of objects. I don't want to read every file when the UI is initalized.
My other option is to mutate the current view model object. Swift does not make this easy
as annoyingly it does not expose the Decoder in JSON decoder, so I had to read a new object
from disk and copy the properties across.

This is where it got extra spicy. I didn't want to write out boiler plate to copy all the properties
across so instead I switched the ViewModel to be an objc object with all objc dynamic members
so I could use KVO to dynamically pull the properties over. This will probably melt the head
of static typing fans, and there are one or two more hacks piled in there for good measures.
Surprisingly it works with SwiftUI